o/i/a/requestrules: store expanded path patterns for rules
When adding a rule to the rule storage, expand the path pattern into a
list of patterns which do not contain groups. Then, try to add each to
the relevant path map, and if there is an identical expanded path
already, then there is a conflicting rule.

Additionally, `RefreshTreeEnforceConsistency` now performs validation on
the constraints, outcome, lifespan, and expiration of the rule, and if
it is invalid, discards it. If a rule with lifespan timespan has an
expiration time in the past, discard it as well. Then, the path pattern
for the rule is expanded and attached to the rule contents as a
non-public non-json field. `RefreshTreeEnforceConsistency` should only
be called when loading rules from disk (where validation ought to
occur), so this does not impose an undue performance penalty during
usual operation.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>