o/i/a/common: adjust patch pattern matching with doublestar
Doublestar v4.6.0 had a bug where paths ending in `/**/` would not match
paths correctly. Doublestar v4.6.1 fixes this bug, but now allows paths
like `/path/to/a/**/` to match `/path/to/a/foo`. We want to restrict
matches further so that patterns ending in `/` only match paths which
also end in `/`. However, we do want patterns which don't end in `/` to
match paths which end in `/`, since received paths which are directories
end in `/` and we don't want to require that the user ends their pattern
in `/` or `{,/}` to match directories or {non,}directories,
respectively.

This, computing whether a pattern matches a path involves three steps:
1. If the pattern ends in `/` but the path doesn't, return false. This
   is really just to catch the case of patterns ending in `/**/`.
2. If the pattern matches the path, return true.
3. Return whether `pattern+'/'` matches the path. This ensures that
   patterns which do not end in `/` match paths which do end in `/`.
   For example, this makes sure that the pattern `/path/to/a` matches
   the path `/path/to/a/`.

Adds some more tests to clarify that this is the case.

Additionally, since `StripTrailingSlashes` is no longer used, remove it,
along with its tests.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>