i/prompting/requestprompts: add requestprompts package
The requestprompts package defines the Prompt structure, which contains
information about a request for which the user should be prompted, and
the PromptDB structure, which contains all unresolved prompts.

Each prompt should must have a unique ID, so a monotonically-increasing
64-bit integer is encoded in hexadecimal with 0-padding and used as the
ID.

If multiple identical requests are received from the kernel, those
listener requests are attached to the Prompt created when the first of
those requests was received. When that prompt is resolved, a reply is
sent back to all listener requests associated with the prompt.

There are two ways of resolving a prompt:

1. The prompt may be replied to directly using the Reply method, which
   sends back a decision to all listener requests associated with the
   prompt.

2. A new rule, either created directly or as a result of a reply with a
   lifespan other than "single", may satisfy a prompt if it either
   has an "allow" outcome and matches *all* remaining permissions, or
   if it has a "deny" outcome and matches *any* remaining permissions.

When a prompt is resolved, or if some but not all of its permissions are
matched by a reply or rule with an "allow" outcome, a new notice is
recorded via the `notifyPrompt` function which is passed in by the
creator of the prompt DB.

When snapd is restarted, the listener is restarted as well, and all
outstanding listener requests are sent a "deny" response. When snapd
starts up again, the new listener is registered on the notify socket,
and the kernel then decides to re-route new (and possibly any old
remaining) requests to the new listener. Thus, there is no reason to
preserve prompts across restarts of snapd, as their backing listener
requests no longer exist.

We would prefer not to re-use prompt IDs across snapd reboots. If a
client attempts to reply to a prompt which no longer exists, it will
receive an error saying the prompt does not exist. However, if a new
prompt with the same ID happens to be created before the client replies,
it may attempt to reply to (what it believes to be) the original prompt,
resulting in unexpected behavior or errors. This commit does not add a
mechanism to ensure that prompt IDs are not re-used across snapd
restarts.

To mitigate this risk somewhat, a `CleanUp` method could be added to the
prompt DB which sends a notice for all outstanding prompts and then
deletes them. The listener already sends "deny" responses for all
outstanding listener requests when `listener.Close` is called, so it is
to be decided whether that responsibility should be shifted to the
prompt DB or left as-is. This has yet to be decided or implemented.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>