i/prompting/requestprompts: make prompt IDs unique across snapd restarts
Previously, if a new prompt with the same ID as another prompt which has
since been resolved happened to be created before a particular prompt
client replied to it, that prompt client might attempt to reply to (what
it believed to be) the original prompt, resulting in unexpected behavior
or errors.

Since prompt IDs are monotonically-increasing integers, we can write the
maximum prompt ID to disk and re-read it whenever snapd starts up. This
ensures that prompt IDs are unique across snapd restarts.

We use a file in `/tmp` to store the maximum ID. Since prompt clients
are restarted when the host system reboots, there is no need to persist
prompt ID uniqueness across system reboots.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>