i/prompting/requestprompts: add clean up function
The new CleanUp function records a notice for every outstanding prompt,
and then deletes all prompts by clearing the per-user prompt DB. This
should be called when snapd is shutting down or restarting.

Note: this does not send a reply to the associated listener requests.
That is left as the responsibility of the listener itself, and the
caller of the CleanUp function should also call `listener.Close()`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>